refactor: restructure directories for better separation of concerns
