+++ backend/chidic-domain/src/test/kotlin/com/domain/service/UserServiceImplTest.kt
@@ -6,7 +6,7 @@ import com.chidicdomain.domain.entity.User
 import com.chidiccommon.enum.Role
 import com.chidicdomain.domain.mapper.UserMapper
 import com.chidicdomain.domain.repository.UserRepository
-import com.chidicdomain.domain.service.UserServiceImpl
+import com.chidicdomain.domain.service.user.UserServiceImpl
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.Mock
